revert 7fffa1e7ea3be2ab1dd80ca3508a49187d3eccea
This introduced an ambiguity for e.g.: return f(x)

Is it to be parsed as [Return(), FunctionCall(...)], or
[Return(FunctionCall(...))]?

I didn't find a good way to disambiguate this.

Compared to pre-7fffa1e7ea3be2ab1dd80ca3508a49187d3eccea,
a recover production has been added allowing the parser to continue
after an incorrectly placed break or return statement.